requirements.txt
langchain
langchain-google-genai
langchain-pinecone
langchain-community
pypdf
google-generativeai
python-dotenv